userns: Change unprivileged_userns_clone proc_handler
Previous unprivileged_userns_clone sysctl proc_handler was proc_dointvec
and allowed the sysctl to be set between 0 and `sizeof(int)`.

Since this sysctl knob is only used to allow or deny unprivileged userns
clone, it should only be set between 0 and 1. Thus, the new proc_handler
is proc_dointvec_minmax with an allowed range value between 0 and 1.

Fixes: ("userns: add sysctl to disallow unprivileged CLONE_NEWUSER by default")
Signed-off-by: Nicolas Bouchinet <nicolas.bouchinet@ssi.gouv.fr>