mm/slub: Only instrument slab allocation with canaries
With barn and sheaves introduction, slab objects are used to prefill or
refill sheaves, which are cache of small objects taking the form of an
array of pointers to slab objects.

Sheaves are then used for quick allocation and free, which consist of
shrinking and growing the array index.
Thus, there is two vision of allocation state for those objects. While
they are seen as allocated by the slab allocator, the sheaf allocator
see them as free and then allocates them.

We thus need to adapt the slab canary patch in order to avoid sanitizing
objects allocation and free from this array.

A next patch will add a per-sheave canary random value which would lead
to a better tracking of objects overflow.

Needs-review-from: Levente Polyak <levente@leventepolyak.net>
Signed-off-by: Nicolas Bouchinet <nicolas.bouchinet@ssi.gouv.fr>